Fake code insight source

+++ myicpc-service-scoreboard/src/main/java/com/myicpc/service/scoreboard/dto/insight/CodeInsightSnapshot.java
@@ -7,15 +7,15 @@ import java.util.Map;
  * @author Roman Smetana
  */
 public class CodeInsightSnapshot implements Serializable {
-    private final int time;
+    private final long time;
     private final Map<Long, CodeInsightProblem> problems;
 
-    public CodeInsightSnapshot(int time, Map<Long, CodeInsightProblem> problems) {
+    public CodeInsightSnapshot(long time, Map<Long, CodeInsightProblem> problems) {
         this.time = time;
         this.problems = problems;
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
     }
